Update README.md
Throughout the project, detailed documentation, code, and visualizations. The documentation serves as an educational resource for individuals interested in machine learning, pattern recognition, and image classification. By combining KNN, ANN, PCA, and BSIF in a structured and systematic manner, the project offers valuable insights into optimizing classifier performance and understanding the role of preprocessing and feature extraction in classification tasks.